+++ macos/Flutter/ephemeral/flutter_export_environment.sh
@@ -1,11 +1,11 @@
 #!/bin/sh
 # This is a generated file; do not edit or check into version control.
-export "FLUTTER_ROOT=/Users/52ndsolution/downloads/development/flutter"
+export "FLUTTER_ROOT=/Users/52ndsolution/downloads/developments/flutter"
 export "FLUTTER_APPLICATION_PATH=/Users/52ndsolution/Desktop/my_packages/tbib_dio_extension"
 export "COCOAPODS_PARALLEL_CODE_SIGN=true"
 export "FLUTTER_BUILD_DIR=build"
-export "FLUTTER_BUILD_NAME=0.0.1"
-export "FLUTTER_BUILD_NUMBER=0.0.1"
+export "FLUTTER_BUILD_NAME=0.0.5"
+export "FLUTTER_BUILD_NUMBER=0.0.5"
 export "DART_OBFUSCATION=false"
 export "TRACK_WIDGET_CREATION=true"
 export "TREE_SHAKE_ICONS=false"